requirements.txt
requests
python-dotenv
paho-mqtt
pandas
xarray
h5netcdf
matplotlib
kafka-python
pyspark